Fix daemon process using CPU when no track is playing
My previous solution to using the Sink object from Rodio was causing the CPU to be used when no tracks were playing. This was due to the Sink keeping the OutputStream open, which prevented the daemon process from staying asleep.

Sinks are now created each time a new track/list is played and destroyed when the track/list has finished. This fix now prevents the daemon process from using unnecessary CPU time when it doesn’t have any tracks to play.